Create Four Sum
The given array is not sorted. The given array may or may not contain duplicate elements. Then take the target as an integer input. Print all the unique quadruple whose sum is equal target.

NOTE all quadruple should be unique, for example : [6, 7, 8, 9], [7, 6, 8, 9] are considered as same quadruple. Also if the array has repeated elements then return only unique quadruple, for eg : if array is arr = [3, 3, 5, 5, 1, 1, 2, 2], and the target = 11, then result will have only one quadruple, i.e. [1, 2, 3, 5].The result should be sorted in increasing order and also the quadruple.