Create Majority_Element
In this task you are supposed to find the majority element of a given array. A majority element of a given array of size N is defined as the element that occurs atleast N/2 times in it.

If N is odd, apply the ceiling function on N/2. So, given an array, compute its majority element.